button controller class working for tap
hold will require an internal pointer to the holding status of a
buttonFns.button array. multi-button buttonFns.button arrays are still
finicky, and do not always trigger (sometimes the single button press is
shown) and so we need to find a way also to limit singles when multiples
are pressed.